+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Add a pop up component - [#30](https://github.com/PrefectHQ/miter-design/pull/30)
+- Add TagGroup component - [#24](https://github.com/PrefectHQ/miter-design/pull/24)
 - Add storybook build command - [#28](https://github.com/PrefectHQ/miter-design/pull/28)
 - Fix miter mixin for specific browsers - [#27](https://github.com/PrefectHQ/miter-design/pull/27)
 - Update miter mixin to make it easier - [#23](https://github.com/PrefectHQ/miter-design/pull/23)